Handle multiple approval request
- fist request comes in as wallet_requestPermissions
- upon approval we then handle the wallet_installSnap request
