+++ app/components/Approvals/InstallSnapApproval/InstallSnapApproval.tsx
@@ -1,82 +1,111 @@
-import React, { useCallback, useState } from 'react';
-import { InstallSnapApprovalFlow } from '../../UI/InstallSnapApprovalFlow';
+import React, { useEffect, useState } from 'react';
 import ApprovalModal from '../ApprovalModal';
 import useApprovalRequest from '../../hooks/useApprovalRequest';
 import { ApprovalTypes } from '../../../core/RPCMethods/RPCMethodMiddleware';
-import Button, {
-  ButtonVariants,
-} from '../../../component-library/components/Buttons/Button';
-import Text from '../../../component-library/components/Texts/Text';
-import { View } from 'react-native';
+import Logger from '../../../util/Logger';
+import { SNAP_INSTALL_FLOW } from '../../../constants/test-ids';
+import { SnapInstallState } from './InstallSnapApproval.types';
+import {
+  InstallSnapConnectionRequest,
+  InstallSnapError,
+  InstallSnapPermissionsRequest,
+  InstallSnapSuccess,
+} from './components';
 
 const InstallSnapApproval = () => {
+  const [installState, setInstallState] = useState<
+    SnapInstallState | undefined
+  >(undefined);
   const [isFinished, setIsFinished] = useState<boolean>(false);
+  const [installError, setInstallError] = useState<Error | undefined>(
+    undefined,
+  );
   const { approvalRequest, onConfirm, onReject } = useApprovalRequest();
 
-  // console.log('SNAPS/ approvalRequest', JSON.stringify(approvalRequest));
+  useEffect(() => {
+    if (approvalRequest) {
+      if (approvalRequest.type === ApprovalTypes.REQUEST_PERMISSIONS) {
+        setInstallState(SnapInstallState.Confirm);
+      } else if (
+        approvalRequest.type === ApprovalTypes.INSTALL_SNAP &&
+        approvalRequest.requestState.permissions
+      ) {
+        setInstallState(SnapInstallState.AcceptPermissions);
+      }
+    } else {
+      setInstallState(undefined);
+    }
+  }, [approvalRequest]);
 
   const onInstallSnapFinished = () => {
     setIsFinished(true);
   };
 
-  // const isWalletRequestPermissions =
-  //   approvalRequest.type === ApprovalTypes.REQUEST_PERMISSIONS &&
-  //   approvalRequest.requestData?.snapId;
+  const onPermissionsConfirm = async () => {
+    try {
+      await onConfirm(undefined, {
+        ...approvalRequest.requestData,
+        permissions: approvalRequest.requestState.permissions,
+      });
+      setInstallState(SnapInstallState.SnapInstalled);
+    } catch (error) {
+      Logger.error(
+        error as Error,
+        `${SNAP_INSTALL_FLOW} Failed to install snap`,
+      );
+      setInstallError(error as Error);
+      setInstallState(SnapInstallState.SnapInstallError);
+    }
+  };
 
-  // const onPermissionsConfirm = useCallback(() => {
-  //   try {
-  //     onConfirm(undefined, {
-  //       ...requestData.requestData,
-  //       permissions: requestData.requestState.permissions,
-  //     });
-  //   } catch (error) {
-  //     Logger.error(
-  //       error as Error,
-  //       `${SNAP_INSTALL_FLOW} Failed to install snap`,
-  //     );
-  //     setInstallError(error as Error);
-  //     setInstallState(SnapInstallState.SnapInstallError);
-  //   }
-  //   setInstallState(SnapInstallState.SnapInstalled);
-  // }, [onConfirm, requestData]);
+  if (!approvalRequest) return null;
 
-  const isWalletRequestPermissions =
-    approvalRequest?.type === ApprovalTypes.REQUEST_PERMISSIONS;
-
-  const onPermissionsConfirm = useCallback(() => {
-    onConfirm(undefined, {
-      ...approvalRequest?.requestData.requestData,
-      permissions: approvalRequest?.requestData.requestState.permissions,
-    });
-  }, []);
-
-  if (isWalletRequestPermissions) {
-    return (
-      <ApprovalModal isVisible={isWalletRequestPermissions} onCancel={onReject}>
-        <View style={{ flex: 1 }}>
-          <Text>This is a permission request</Text>
-          <Button
-            variant={ButtonVariants.Primary}
-            onPress={() => onPermissionsConfirm}
-            label="confirm"
+  const renderModalContent = () => {
+    switch (installState) {
+      case SnapInstallState.Confirm:
+        return (
+          <InstallSnapConnectionRequest
+            approvalRequest={approvalRequest}
+            requestState={approvalRequest.requestState}
+            onConfirm={onConfirm}
+            onCancel={onReject}
           />
-        </View>
-      </ApprovalModal>
-    );
-  }
-
-  const isModalVisible: boolean =
-    approvalRequest?.requestData.type === ApprovalTypes.INSTALL_SNAP ||
-    (!isFinished && approvalRequest !== undefined);
+        );
+      case SnapInstallState.AcceptPermissions:
+        return (
+          <InstallSnapPermissionsRequest
+            approvalRequest={approvalRequest}
+            requestState={approvalRequest.requestState}
+            onConfirm={onPermissionsConfirm}
+            onCancel={onReject}
+          />
+        );
+      case SnapInstallState.SnapInstalled:
+        return (
+          <InstallSnapSuccess
+            approvalRequest={approvalRequest}
+            onConfirm={onInstallSnapFinished}
+          />
+        );
+      case SnapInstallState.SnapInstallError:
+        return (
+          <InstallSnapError
+            approvalRequest={approvalRequest}
+            onConfirm={onInstallSnapFinished}
+            error={installError}
+          />
+        );
+      default:
+        return null;
+    }
+  };
 
   return (
-    <ApprovalModal isVisible={isModalVisible} onCancel={onReject}>
-      <InstallSnapApprovalFlow
-        onCancel={onReject}
-        onConfirm={onConfirm}
-        onFinish={onInstallSnapFinished}
-        requestData={approvalRequest}
-      />
+    <ApprovalModal
+      isVisible={installState !== undefined && !isFinished}
+      onCancel={onReject}
+    >
+      {renderModalContent()}
     </ApprovalModal>
   );
 };
