+++ app/components/Approvals/InstallSnapApproval/components/InstallSnapConnectionRequest/index.ts
