+++ app/components/Approvals/InstallSnapApproval/components/InstallSnapConnectionRequest/test/InstallSnapConnectionRequest.test.tsx
