+++ app/components/Approvals/InstallSnapApproval/components/InstallSnapError/InstallSnapError.styles.ts
