+++ app/components/Approvals/InstallSnapApproval/components/InstallSnapError/index.ts
