+++ app/components/Approvals/InstallSnapApproval/components/InstallSnapError/test/InstallSnapError.test.tsx
