+++ app/components/Approvals/InstallSnapApproval/components/InstallSnapPermissionsRequest/InstallSnapPermissionsRequest.tsx
@@ -1,19 +1,16 @@
 import React, { useMemo } from 'react';
-import { ImageSourcePropType, View } from 'react-native';
-import { InstallSnapFlowProps } from '../../InstallSnapApprovalFlow.types';
-import styleSheet from './InstallSnapConnectionRequest.styles';
+import { ScrollView, View } from 'react-native';
+import stylesheet from './InstallSnapPermissionRequest.styles';
 import { strings } from '../../../../../../locales/i18n';
 import {
   SNAP_INSTALL_CANCEL,
-  SNAP_INSTALL_CONNECT,
-  SNAP_INSTALL_CONNECTION_REQUEST,
+  SNAP_INSTALL_PERMISSIONS_REQUEST,
+  SNAP_INSTALL_PERMISSIONS_REQUEST_APPROVE,
 } from '../../../../../constants/test-ids';
 import SheetHeader from '../../../../../component-library/components/Sheet/SheetHeader';
 import Text, {
   TextVariant,
 } from '../../../../../component-library/components/Texts/Text';
-import TagUrl from '../../../../../component-library/components/Tags/TagUrl';
-import { getUrlObj, prefixUrlWithProtocol } from '../../../../../util/browser';
 import { IconName } from '../../../../../component-library/components/Icons/Icon';
 import Cell, {
   CellVariant,
@@ -28,50 +25,27 @@ import BottomSheetFooter, {
 } from '../../../../../component-library/components/BottomSheets/BottomSheetFooter';
 import { ButtonProps } from '../../../../../component-library/components/Buttons/Button/Button.types';
 import { useStyles } from '../../../../hooks/useStyles';
+import { InstallSnapFlowProps } from '../../InstallSnapApproval.types';
+import { SnapPermissions } from '../../../../Views/Snaps/components/SnapPermissions';
 
-const InstallSnapConnectionRequest = ({
-  requestData,
+const InstallSnapPermissionsRequest = ({
+  approvalRequest,
   requestState,
   onConfirm,
   onCancel,
 }: InstallSnapFlowProps) => {
-  const { styles } = useStyles(styleSheet, {});
-  console.log(
-    'SNAPS/ InstallSnapConnectionRequest.tsx/ requestData: ',
-    JSON.stringify(requestData, null, 2),
-  );
-
-  console.log(
-    'SNAPS/ InstallSnapConnectionRequest.tsx/ requestState: ',
-    JSON.stringify(requestState, null, 2),
-  );
-
+  const { styles } = useStyles(stylesheet, {});
   const snapName = useMemo(() => {
-    const colonIndex = requestData.snapId.indexOf(':');
+    const colonIndex = approvalRequest.requestData.snapId.indexOf(':');
     if (colonIndex !== -1) {
-      return requestData.snapId.substring(colonIndex + 1);
+      return approvalRequest.requestData.snapId.substring(colonIndex + 1);
     }
-    return requestData.snapId;
-  }, [requestData.snapId]);
+    return approvalRequest.requestData.snapId;
+  }, [approvalRequest.requestData.snapId]);
 
-  const dappOrigin = useMemo(
-    () => requestData.metadata.dappOrigin,
-    [requestData.metadata.dappOrigin],
-  );
-
-  const favicon: ImageSourcePropType = useMemo(() => {
-    const iconUrl = `https://api.faviconkit.com/${dappOrigin}/50`;
-    return { uri: iconUrl };
-  }, [dappOrigin]);
-
-  const urlWithProtocol = prefixUrlWithProtocol(dappOrigin);
-
-  const secureIcon = useMemo(
-    () =>
-      (getUrlObj(dappOrigin) as URL).protocol === 'https:'
-        ? IconName.Lock
-        : IconName.LockSlash,
-    [dappOrigin],
+  const origin = useMemo(
+    () => approvalRequest.origin,
+    [approvalRequest.origin],
   );
 
   const cancelButtonProps: ButtonProps = {
@@ -84,27 +58,15 @@ const InstallSnapConnectionRequest = ({
 
   const connectButtonProps: ButtonProps = {
     variant: ButtonVariants.Primary,
-    label: strings('accountApproval.connect'),
+    label: strings('install_snap.approve_permissions'),
     size: ButtonSize.Lg,
     onPress: onConfirm,
-    testID: SNAP_INSTALL_CONNECT,
+    testID: SNAP_INSTALL_PERMISSIONS_REQUEST_APPROVE,
   };
 
   return (
-    <View testID={SNAP_INSTALL_CONNECTION_REQUEST} style={styles.root}>
+    <View testID={SNAP_INSTALL_PERMISSIONS_REQUEST} style={styles.root}>
       <View style={styles.accountCardWrapper}>
-        <TagUrl
-          imageSource={favicon}
-          label={urlWithProtocol}
-          iconName={secureIcon}
-        />
-        <SheetHeader title={strings('install_snap.title')} />
-        <Text style={styles.description} variant={TextVariant.BodyMD}>
-          {strings('install_snap.description', {
-            origin: dappOrigin,
-            snap: snapName,
-          })}
-        </Text>
         <Cell
           style={styles.snapCell}
           variant={CellVariant.Display}
@@ -114,6 +76,21 @@ const InstallSnapConnectionRequest = ({
             name: IconName.Snaps,
           }}
         />
+        <SheetHeader
+          title={strings('install_snap.permissions_request_title')}
+        />
+        <Text style={styles.description} variant={TextVariant.BodyMD}>
+          {strings('install_snap.permissions_request_description', {
+            origin,
+            snap: snapName,
+          })}
+        </Text>
+        <ScrollView style={styles.snapPermissionContainer}>
+          <SnapPermissions
+            permissions={requestState.permissions}
+            showLabel={false}
+          />
+        </ScrollView>
         <View style={styles.actionContainer}>
           <BottomSheetFooter
             buttonsAlignment={ButtonsAlignment.Horizontal}
@@ -125,4 +102,4 @@ const InstallSnapConnectionRequest = ({
   );
 };
 
-export default InstallSnapConnectionRequest;
+export default InstallSnapPermissionsRequest;
