+++ app/components/Approvals/InstallSnapApproval/components/InstallSnapPermissionsRequest/index.ts
