+++ app/components/Approvals/InstallSnapApproval/components/InstallSnapPermissionsRequest/test/InstallSnapPermissionsRequest.test.tsx
