+++ app/components/Approvals/InstallSnapApproval/components/InstallSnapSuccess/InstallSnapSuccess.styles.ts
