+++ app/components/Approvals/InstallSnapApproval/components/InstallSnapSuccess/index.ts
