+++ app/components/Approvals/InstallSnapApproval/components/InstallSnapSuccess/test/InstallSnapSuccess.test.tsx
