+++ app/components/Approvals/InstallSnapApproval/components/index.ts
@@ -1,0 +1,9 @@
+import { InstallSnapConnectionRequest } from './InstallSnapConnectionRequest';
+import { InstallSnapSuccess } from './InstallSnapSuccess';
+import { InstallSnapError } from './InstallSnapError';
+import { InstallSnapPermissionsRequest } from './InstallSnapPermissionsRequest';
+
+export { InstallSnapPermissionsRequest };
+export { InstallSnapError };
+export { InstallSnapSuccess };
+export { InstallSnapConnectionRequest };
