+++ app/components/Approvals/InstallSnapApproval/test/InstallSnapApproval.test.tsx
