--- app/components/UI/InstallSnapApprovalFlow/index.ts
@@ -1,5 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-import { InstallSnapApprovalArgs } from './InstallSnapApprovalFlow.types';
-import InstallSnapApprovalFlow from './InstallSnapApprovalFlow';
-export { InstallSnapApprovalFlow };
-export type { InstallSnapApprovalArgs };
