+++ app/components/hooks/useApprovalRequest.ts
@@ -13,11 +13,6 @@ const useApprovalRequest = () => {
     | ApprovalRequest<any>
     | undefined;
 
-  console.log(
-    'SNAPS/ useApprovalRequest.ts: pendingApprovals: ',
-    JSON.stringify(pendingApprovals),
-  );
-
   const pageMeta = approvalRequest?.requestData?.pageMeta ?? {};
 
   const onConfirm = useCallback(
@@ -25,7 +20,6 @@ const useApprovalRequest = () => {
       opts?: Parameters<typeof Engine.acceptPendingApproval>[2],
       value?: Parameters<typeof Engine.acceptPendingApproval>[1],
     ) => {
-      console.log('SNAPS/ useApprovalRequest.ts: onConfirm:');
       if (!approvalRequest) return;
 
       await Engine.acceptPendingApproval(
@@ -45,11 +39,6 @@ const useApprovalRequest = () => {
       ethErrors.provider.userRejectedRequest(),
     );
   }, [approvalRequest]);
-
-  console.log(
-    'SNAPS/ useApprovalRequest.ts: pageMeta: ',
-    JSON.stringify(pageMeta, null, 2),
-  );
 
   return {
     approvalRequest: cloneDeep(approvalRequest),
