+++ app/core/Snaps/location/npm.ts
@@ -75,7 +75,6 @@ const decompressFile = async (
   path: string,
   targetPath: string,
 ): Promise<string> => {
-  console.log('SNAPS/ NPM/ decompressFile', path, targetPath);
   try {
     const decompressedDataLocation = await RNTar.unTar(path, targetPath);
     if (decompressedDataLocation) {
@@ -98,7 +97,6 @@ const readAndParseAt = async (path: string) => {
 const fetchAndStoreNPMPackage = async (
   inputRequest: RequestInfo,
 ): Promise<string> => {
-  console.log('SNAPS/ NPM/ fetchAndStoreNPMPackage', inputRequest);
   const { config } = RNFetchBlob;
   const targetDir = RNFetchBlob.fs.dirs.DocumentDir;
   const filePath = `${targetDir}/archive.tgz`;
@@ -156,7 +154,6 @@ export class NpmLocation implements SnapLocation {
   private files?: Map<string, VirtualFile>;
 
   constructor(url: URL, opts: DetectSnapLocationOptions = {}) {
-    console.log('SNAPS/ NpmLocation consructor', url, opts);
     const allowCustomRegistries = opts.allowCustomRegistries ?? false;
     const fetchFunction = opts.fetch ?? globalThis.fetch.bind(globalThis);
     const requestedRange = opts.versionRange ?? DEFAULT_REQUESTED_SNAP_VERSION;
@@ -270,8 +267,6 @@ export class NpmLocation implements SnapLocation {
         this.meta.fetch,
       );
     this.meta.version = actualVersion;
-
-    console.log('SNAPS/ NPM/ #lazyInit', manifestData, pathToFiles);
 
     let canonicalBase = 'npm://';
     if (this.meta.registry.username !== '') {
